requirements.txt
discord.py[voice]
requests
py_dotenv
animalapi
psutil
mysql